requirements.txt
gspread=5.12.4
oauth2client=4.1.3
pandas=2.1.4
psycopg2-binary=2.9.9
pymongo=4.6.1
python-dotenv=1.0.0
